split declaration and impl
